requirements.txt
asyncio
discord.py
beautifulsoup4
psycopg2
lxml
aiohttp
websockets
chardet